ocr-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
easyocr==1.7.2
Pillow==11.0.0
numpy==1.26.4
python-dotenv==1.0.1
